Update Braille-display-code-2019
This code has a representation of Braille characters in a 3*2 list of lists. The Picam module of the Raspberry pi captures the image which is then extracted of texts using the OCR library Pytesseract. The extracted text is then iterated letter by letter and displayed in a Braille format with Servo motors arranged in 3*2 formation. 
This is the code that I wrote from scratch back in 2019.